fix(extras): allow SET_ZERO_REFERENCE_POSITION to be called without having a beacon
- Fixes #124
- Moved the implementation from a beacon-specific module to the general-purpose [ratos] module
